+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Feat: Add delete network graph in settings
+- Fix: Make switch buttons in receive screen not jumpy
 
 ## [1.7.2] - 2023-12-13